Enable addon usage on private TBC servers
The "GetCategoryList" is a WoW 3.x+ API which gets a list of achievement
categories. In the past, QuestHelper used it to detect private servers.

But it makes no sense on 2.4.3 TBC, since that API doesn't exist in TBC
clients, and was therefore preventing QuestHelper from loading.

By simply removing that check, the addon now loads in TBC.