Remove "new" data collection system
QuestHelper's author had been working on a "new" data collection
system which grabbed information about NPCs, quests, loot, etc, as you
walk around the world and play the game. This data is then submitted to
the author (manually) to update the database. But since this version of
QuestHelper will never be updated (and doesn't need database updates),
it's much more efficient to just rip out and remove all of that "new"
data collection code.

This will increase the FPS and general performance, since there's no
need to maintain that "new" data collection database every time you do
something in-game.

There's still an "old" data collector, in the various files described in
the TOC file. It seems to still be active, but is too deeply embedded to
touch. It doesn't matter as much.