Fix Stormwind & Eastern Plaguelands TBC routing
The database comes from Wrath, after Blizzard redesigned Stormwind
City and Eastern Plaguelands and moved all of the objectives/npcs.

This means that the database is incorrect in those zones in TBC!

We therefore need to translate coordinates for any objective (quest,
item, npc, flightmaster, zone border, etc) in those zones, to restore
their locations back to their pre-Wrath "TBC" positions again, so
that the map icons and arrows etc point the user to the correct place!

We achieve this by performing the translation in the Objective's
"AddLoc()" function, which is the function where all locations
are added to an objective and converted to their final map-offsets,
which means that it's the *perfect* place for inserting this fix.

The translation happens at runtime, and ONLY for *relevant* objectives
that YOUR character needs. So there's NO extra overhead! :-)

This concludes the work on backporting QuestHelper from Wrath to TBC!

Enjoy! :-)